extract getLatestVersion from packagist-version to helper fn
no functional changes here - just moving code around
